backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36

# Auth / JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client (for Supabase REST calls)
httpx==0.28.1
supabase==2.12.0

# Image processing / OCR
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
Pillow==11.1.0
numpy==2.0.2
scikit-image==0.24.0

# Logging
python-json-logger==3.2.1

# Utilities
python-multipart==0.0.20
python-dotenv==1.0.1

# Reporting
weasyprint==62.3
Jinja2==3.1.4

# PDF processing
pdf2image==1.17.0
pypdfium2==4.30.0
